requirements.txt
flask==2.3.1
pytest==7.3.1
selenium==4.9.0
ujson==5.7.0
wikipedia-api==0.5.8
psycopg2-binary==2.9.6
flask-sock